Domain, Data Layer 계층 모듈화
