+++ DataLayerTests/DataTests.swift
@@ -1,0 +1,17 @@
+//
+//  DataTests.swift
+//  DataTests
+//
+//  Created by 조영태 on 7/18/25.
+//
+
+import Testing
+@testable import Data
+
+struct DataTests {
+
+    @Test func example() async throws {
+        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+    }
+
+}
