+++ Domain/Domain.h
@@ -1,0 +1,18 @@
+//
+//  Domain.h
+//  Domain
+//
+//  Created by 조영태 on 7/18/25.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for Domain.
+FOUNDATION_EXPORT double DomainVersionNumber;
+
+//! Project version string for Domain.
+FOUNDATION_EXPORT const unsigned char DomainVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <Domain/PublicHeader.h>
+
+
