+++ DomainTests/DomainTests.swift
@@ -1,0 +1,17 @@
+//
+//  DomainTests.swift
+//  DomainTests
+//
+//  Created by 조영태 on 7/18/25.
+//
+
+import Testing
+@testable import Domain
+
+struct DomainTests {
+
+    @Test func example() async throws {
+        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+    }
+
+}
