+++ todoApp/DI/Container/UIKit/TodoListDIContainer.UIK.swift
@@ -6,9 +6,10 @@
 //
 
 import Foundation
-import Moya
 import Swinject
 import UIKit
+import Domain
+import DataLayer
 
 extension UIK {
     class TodoListDIContainer {
