+++ todoApp/Data/RelamModel/TodoRealm.swift
@@ -1,28 +1,28 @@
+////
+////  TodoRealm.swift
+////  todoApp
+////
+////  Created by 조영태 on 3/11/25.
+////
 //
-//  TodoRealm.swift
-//  todoApp
+//import Foundation
+//import RealmSwift
 //
-//  Created by 조영태 on 3/11/25.
 //
-
-import Foundation
-import RealmSwift
-
-
-class TodoRealm: Object {
-    @Persisted(primaryKey: true) var _id: String
-    @Persisted var title: String
-    @Persisted var date: Date
-    @Persisted var contents: String
-    @Persisted var isDone: Bool
-
-    
-    convenience init(title: String, date: Date, contents: String) {
-        self.init()
-        self._id = UUID().uuidString
-        self.title = title
-        self.date = date
-        self.contents = contents
-        self.isDone = false
-    }
-}
+//class TodoRealm: Object {
+//    @Persisted(primaryKey: true) var _id: String
+//    @Persisted var title: String
+//    @Persisted var date: Date
+//    @Persisted var contents: String
+//    @Persisted var isDone: Bool
+//
+//    
+//    convenience init(title: String, date: Date, contents: String) {
+//        self.init()
+//        self._id = UUID().uuidString
+//        self.title = title
+//        self.date = date
+//        self.contents = contents
+//        self.isDone = false
+//    }
+//}
