+++ todoApp/Domain/Entity/Todo.swift
@@ -7,27 +7,27 @@
 
 import Foundation
 
-protocol CreatableTodo {
-    var title: String { get }
-    var date: Date { get }
-    var contents: String { get }
-}
-
-struct CreatableToDoImpl: CreatableTodo {
-    var title: String
-    var date: Date
-    var contents: String
-}
-
-protocol Todo : CreatableTodo {
-    var id: String { get }
-    var isDone: Bool { get }
-}
-
-struct TodoImpl: Todo {
-    var id: String
-    var title: String
-    var date: Date
-    var contents: String
-    var isDone: Bool
-}
+//protocol CreatableTodo {
+//    var title: String { get }
+//    var date: Date { get }
+//    var contents: String { get }
+//}
+//
+//struct CreatableToDoImpl: CreatableTodo {
+//    var title: String
+//    var date: Date
+//    var contents: String
+//}
+//
+//protocol Todo : CreatableTodo {
+//    var id: String { get }
+//    var isDone: Bool { get }
+//}
+//
+//struct TodoImpl: Todo {
+//    var id: String
+//    var title: String
+//    var date: Date
+//    var contents: String
+//    var isDone: Bool
+//}
