+++ todoApp/Domain/UseCase/EditTodoUseCase.swift
@@ -1,53 +1,53 @@
+////
+////  EditTodoUseCase.swift
+////  todoApp
+////
+////  Created by 조영태 on 5/11/25.
+////
 //
-//  EditTodoUseCase.swift
-//  todoApp
+//import Foundation
 //
-//  Created by 조영태 on 5/11/25.
+//protocol EditTodoUseCase {
+//    associatedtype Error: Swift.Error
+//    var repository: TodoRepository { get }
+//    
+//    /// Todo 수정하기
+//    /// - Throws: ``EditTodoUseCase.Error``
+//    func execute(_ target:Todo, newTitle:String?, newDate:Date?, newContents:String?) async throws -> Todo
+//}
 //
-
-import Foundation
-
-protocol EditTodoUseCase {
-    associatedtype Error: Swift.Error
-    var repository: TodoRepository { get }
-    
-    /// Todo 수정하기
-    /// - Throws: ``EditTodoUseCase.Error``
-    func execute(_ target:Todo, newTitle:String?, newDate:Date?, newContents:String?) async throws -> Todo
-}
-
-final class DefaultEditTodoUseCase: EditTodoUseCase {
-    enum Error: Swift.Error {
-        case notFound
-    }
-
-    var repository: any TodoRepository
-    init(repository: any TodoRepository) {
-        self.repository = repository
-    }
-    
-    func execute(_ target:Todo, newTitle:String?, newDate:Date?, newContents:String? ) async throws -> Todo {
-        do {
-            
-            let newTodo  = TodoImpl(
-                id: target.id,
-                title: newTitle ?? target.title,
-                date: newDate ?? target.date,
-                contents: newContents ?? target.contents,
-                isDone: target.isDone
-            )
-            
-            let response = try await repository.updateTodo(newTodo)
-            
-            return response
-        } catch TodoError.notFound {
-            // DataLayer Error Mapping
-            throw Error.notFound
-        } catch {
-            // Network Error 전달
-            throw error
-        }
-    }
-    
-}
-
+//final class DefaultEditTodoUseCase: EditTodoUseCase {
+//    enum Error: Swift.Error {
+//        case notFound
+//    }
+//
+//    var repository: any TodoRepository
+//    init(repository: any TodoRepository) {
+//        self.repository = repository
+//    }
+//    
+//    func execute(_ target:Todo, newTitle:String?, newDate:Date?, newContents:String? ) async throws -> Todo {
+//        do {
+//            
+//            let newTodo  = TodoImpl(
+//                id: target.id,
+//                title: newTitle ?? target.title,
+//                date: newDate ?? target.date,
+//                contents: newContents ?? target.contents,
+//                isDone: target.isDone
+//            )
+//            
+//            let response = try await repository.updateTodo(newTodo)
+//            
+//            return response
+//        } catch TodoError.notFound {
+//            // DataLayer Error Mapping
+//            throw Error.notFound
+//        } catch {
+//            // Network Error 전달
+//            throw error
+//        }
+//    }
+//    
+//}
+//
