+++ todoApp/SceneDelegate.swift
@@ -7,6 +7,7 @@
 
 import UIKit
 import RealmSwift
+import DataLayer
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
@@ -25,20 +26,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         
         window?.makeKeyAndVisible()
 //        mainVC.present(ViewController(), animated: false)
-
-        let config = Realm.Configuration(
-            schemaVersion: 2, //
-            migrationBlock: { migration, oldSchemaVersion in
-                if oldSchemaVersion < 2 {
-                    // 예: isDone 프로퍼티가 새로 생겼다면 초기값 세팅도 가능
-                    migration.enumerateObjects(ofType: TodoRealm.className()) { oldObject, newObject in
-                        newObject?["isDone"] = false
-                    }
-                }
-            }
-        )
-
-        Realm.Configuration.defaultConfiguration = config
+        RealmConfig.setConfig()
+        
 #if DEBUG
         print("realm 위치: ", Realm.Configuration.defaultConfiguration.fileURL!)
 #endif
